+++ features/step_definitions/admin_steps/member/verify_with_scra_steps.rb
@@ -10,6 +10,13 @@ Given(/^I should be successfully verified as a "([^"]*)"$/) do |group|
   VerifyWithScra.new.verify_scra_applied(@user_email, affiliation: group)
 end
 
+Given(/^I submit the SCRA verification form through the admin tool as a "([^"]*)" record$/) do |type|
+  VerifyWithScra.new.verify_via_scra(@user_email, affiliation: "Service Member", type: type)
+end
+
+Given(/^I should see the red alert flash message$/) do
+  VerifyWithScra.new.verify_denied_scra_error_message
+
 Given(/^I submit the empty SCRA form for "([^"]*)" in the Admin tool$/) do |group|
   VerifyWithScra.new.verify_via_scra(@user_email, affiliation: group, populate: false)
 end